% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarPageWithInputs.R
\name{navbarPageWithInputs}
\alias{navbarPageWithInputs}
\title{Navigation Bar with Input Widgets}
\usage{
navbarPageWithInputs(..., input1, input2)
}
\arguments{
\item{...}{Arguments in shiny::navbarPage}

\item{input1}{first input widget}

\item{input2}{second input widget}
}
\value{
A navigation bar with input widgets within it
}
\description{
Navigation Bar with Input Widgets
}
